Restructuring: renamed project and server packages
