+++ pom.xml
@@ -3,14 +3,16 @@
     xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-	<groupId>com.example</groupId>
-    <artifactId>demo</artifactId>
+	<groupId>dsv.pvt2018</groupId>
+    <artifactId>CaptureTheLamp</artifactId>
     <version>0.1.0</version>
     <packaging>pom</packaging>
 
     <modules>
         <module>server</module>
+        <!--
         <module>client</module>
+        -->
     </modules>
 
 </project>
