+++ server/src/main/java/dsv/pvt2018/UserRepository.java
@@ -1,8 +1,6 @@
-package com.example.demo;
+package dsv.pvt2018;
 
 import org.springframework.data.repository.CrudRepository;
-
-import com.example.demo.User;
 
 // This will be AUTO IMPLEMENTED by Spring into a Bean called userRepository
 // CRUD refers Create, Read, Update, Delete
